dhanalakshmi5.1/requirements.txt
flask
psycopg2-binary
reportlab
werkzeug
python-dotenv
scikit-learn
